+++ Kmandili/Kmandili/Models/RestClient/UploadRestClient.cs
@@ -2,8 +2,10 @@
 using System.Collections.Generic;
 using System.IO;
 using System.Net.Http;
+using System.Net.Http.Headers;
 using System.Text;
 using System.Threading.Tasks;
+using Kmandili.Helpers;
 using Plugin.Connectivity;
 
 namespace Kmandili.Models.RestClient
@@ -30,8 +32,9 @@ namespace Kmandili.Models.RestClient
 
         public async Task<bool> Delete(string fileName)
         {
-            if (!(await CheckConnection())) return false;
+            if (!(await CheckConnection()) || (App.TokenExpired())) return false;
             var httpClient = new HttpClient();
+            httpClient.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", Settings.Token);
 
             string URL = App.ServerURL + "api/Uploads/" + fileName;
             var response = await httpClient.DeleteAsync(URL);
